requirements.txt
alembic==1.4.2
aniso8601==6.0.0
astroid==2.4.1
atomicwrites==1.4.0
attrs==19.3.0
Authlib==0.14.3
autopep8==1.5.2
awscli==1.18.124
Babel==2.7.0
botocore==1.17.47
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
Click==7.0
colorama==0.4.3
console-log==0.2.10
cryptography==3.0
docformatter==1.3.1
docutils==0.15.2
ecdsa==0.15
Flask==1.0.3
Flask-Cors==3.0.7
Flask-Migrate==2.5.3
Flask-Moment==0.9.0
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.3
gevent==20.5.0
gevent-websocket==0.10.1
greenlet==0.4.15
gunicorn==20.0.4
idna==2.10
iniconfig==1.0.1
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.10.0
lazy-object-proxy==1.4.3
Mako==1.1.2
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.4.0
packaging==20.4
phonenumbers==8.12.3
pluggy==0.13.1
psycopg2-binary==2.8.5
py==1.9.0
pyasn1==0.4.8
pycodestyle==2.5.0
pycparser==2.20
PyJWT==1.7.1
pylint==2.5.2
pyparsing==2.4.7
pytest==6.0.1
python-dateutil==2.8.1
python-editor==1.0.4
python-jose==3.1.0
pytz==2019.1
PyYAML==5.3.1
requests==2.24.0
rsa==4.5
s3transfer==0.3.3
six==1.15.0
SQLAlchemy==1.3.17
toml==0.10.0
untokenize==0.1.1
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
wsgigzip==0.1.4
WTForms==2.3.1
